+++ packages/utils/src/string.ts
@@ -10,6 +10,9 @@ export const nonEmptyStr = (x?: any): x is string =>
   isStr(x) && x.trim().length > 0;
 
 export const asString = (x: { toString: () => string }): string => {
-  return typeof x === 'string' ? x : x.toString()
+  return typeof x === 'string' ? x : x?.toString()
 }
-  
+
+export const asStringArray = (x: { toString: () => string }[]): string[] => {
+  return x.map(asString)
+}