+++ CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.8.1 - 2023-02-17
+
+### Added
+- Add flatten function for reactions
+- Add `findReaction` and `findReactions` to top-level API (wrappers around `api.blockchain.findReaction` and `api.blockchain.findReactions`)
+- Add new type `ReactionStruct`
+- Add new attribute `username` for Space in Elastic Index type
+- Add typedoc for `@subsocial/utils` package
+
+### Fixed
+- Fix behavior of `simpleFormatBalance` `currency` argument not working
+
+### Changed
+- Change the return type of ids in `api.blockchain` functions to string
+- Change argument of `createTwitterURL` to accept object with only required properties
+- Make `api.blockchain.getReactionIdsByAccount` private, because its same as `api.blockchain.getPostReactionIdsByAccount`
+
+### Deprecated
+- Make `api.base` deprecated. All the functionality provided by it are all implemented in the top-level API. e.g. `api.base.findPost` <=> `api.findPost`
+- Deprecate `handle` attribute in Elastic Space type. Use newly added field `username` instead
+
+### Removed
+- Remove `profiles` from elastic index types
+
 ## 0.7.13 - 2023-02-13
 ### Added
 - add `get-metadata.js` for definitions, for fast updating types after chain upgrades